+++ packages/pyroscope-flamegraph/CHANGELOG.md
@@ -3,6 +3,81 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.16.0](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/flamegraph@0.15.0...@pyroscope/flamegraph@0.16.0) (2022-07-25)
+
+
+### Features
+
+* show functions % of total [units] in Table ([#1288](https://github.com/pyroscope-io/pyroscope/issues/1288)) ([6c71195](https://github.com/pyroscope-io/pyroscope/commit/6c71195295c3ed5591917ff754e670d4220b77d0))
+
+
+
+
+
+# [0.15.0](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/flamegraph@0.14.9...@pyroscope/flamegraph@0.15.0) (2022-07-25)
+
+
+### chore
+
+* **flamegraph/models:** make it mandatory to handle all spyNames ([#1300](https://github.com/pyroscope-io/pyroscope/issues/1300)) ([f7a95a0](https://github.com/pyroscope-io/pyroscope/commit/f7a95a0225c1a39262962a47fd2a1cd493a8333b))
+
+
+### BREAKING CHANGES
+
+* **flamegraph/models:** it will throw an error if spyName is unsupported
+
+
+
+
+
+## [0.14.9](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/flamegraph@0.14.8...@pyroscope/flamegraph@0.14.9) (2022-07-20)
+
+**Note:** Version bump only for package @pyroscope/flamegraph
+
+
+
+
+
+## [0.14.8](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/flamegraph@0.14.7...@pyroscope/flamegraph@0.14.8) (2022-07-19)
+
+**Note:** Version bump only for package @pyroscope/flamegraph
+
+
+
+
+
+## [0.14.7](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/flamegraph@0.14.6...@pyroscope/flamegraph@0.14.7) (2022-07-19)
+
+**Note:** Version bump only for package @pyroscope/flamegraph
+
+
+
+
+
+## [0.14.6](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/flamegraph@0.14.5...@pyroscope/flamegraph@0.14.6) (2022-07-18)
+
+**Note:** Version bump only for package @pyroscope/flamegraph
+
+
+
+
+
+## [0.14.5](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/flamegraph@0.14.4...@pyroscope/flamegraph@0.14.5) (2022-07-18)
+
+**Note:** Version bump only for package @pyroscope/flamegraph
+
+
+
+
+
+## [0.14.4](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/flamegraph@0.14.3...@pyroscope/flamegraph@0.14.4) (2022-07-18)
+
+**Note:** Version bump only for package @pyroscope/flamegraph
+
+
+
+
+
 ## [0.14.3](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/flamegraph@0.14.2...@pyroscope/flamegraph@0.14.3) (2022-07-15)
 
 **Note:** Version bump only for package @pyroscope/flamegraph